requirements.txt
Django==2.2.17
djangorestframework==3.12.2
django-rest-swagger==2.2.0
mysqlclient==2.0.2
django-mysql==2.2.0
django-phone-field==1.8.1
factory-boy==3.1.0